requirements.txt
Pillow>=9.0.0
numpy>=1.19.0
torch>=2.0.0
torchvision>=0.15.0
safetensors>=0.3.1
groundingdino @ git+https://github.com/IDEA-Research/GroundingDINO.git